Change PHP version and add PDO extension in composer.json
This commit includes the dependency changes in the composer.json file.
The changes include the addition of `ext-pdo` extension requirement and
moving the `barryvdh/laravel-debugbar` package from the list of requirements
to the list of required packages. In addition, there is a change in the
PHP version in the platform section to ^8.2. The composer.lock file has been
updated accordingly.